CalcEx implemente for inelastic and other reactions

+++ Parameters.c
@@ -99,7 +99,7 @@ int *TOFOffsets;
 
 int NrOfRunsForPadOffsets;
 int *RunNrForPadOffsets;
-int *PadOffsets;
+double *PadOffsets;
 
 
 bool TestInelastic = true; //Test to see if this is an elastic reaction... default is true as they're the ones that we run the most
@@ -1285,7 +1285,7 @@ void ReadConfiguration()
 		  NrOfRunsForPadOffsets = atoi(LineBuffer.c_str());
                   if(NrOfRunsForPadOffsets<1) NrOfRunsForPadOffsets=1;    //if you put 0 in config I will create at least 1 entry for safety
                   RunNrForPadOffsets = new int[NrOfRunsForPadOffsets];
-                  PadOffsets = new int[NrOfRunsForPadOffsets];
+                  PadOffsets = new double[NrOfRunsForPadOffsets];
 	        }
 	      else if(LineBuffer.compare(0,14,"PadOffsetsFile") == 0)
 		{
